requirements.txt
pytest
selenium
requests
pytest-html
pytest-xdist
pandas
matplotlib
allure-pytest
pytest-rerunfailures


